check auth before crud

+++ backend/src/app/modules/comment/comment.service.ts
@@ -1,17 +1,77 @@
 import { JwtPayload } from "../../../interfaces/common"
 import { ICommentsByVideoId, ICreateComment, IUpdateComment } from "./comment.interface"
 import prisma from "../../../shared/prisma"
+import ApiError from "../../../errors/ApiError"
+import httpStatus from "http-status"
+
+const checkAccess = async (user: JwtPayload, videoId: string) => {
+    const video = await prisma.video.findUnique({
+        where: { id: videoId },
+        include: {
+            module: {
+                include: {
+                    milestone: {
+                        include: {
+                            course: true
+                        }
+                    }
+                }
+            }
+        }
+
+    },)
+    const instructorId = (video?.module.milestone.course.instructorId)
+    const courseId = (video?.module.milestone.course.id)
+    if (user.role === 'INSTRUCTOR' && user.userId === instructorId) {
+        return true
+    }
+
+    const isCourseBought = await prisma.transactions.findFirst({
+        where: {
+            AND: [
+                {
+                    studentId: {
+                        equals: user.userId
+                    }
+                },
+                {
+                    courseId: {
+                        equals: courseId
+                    }
+                }
+            ]
+
+
+        }
+    })
+    if (isCourseBought) {
+        return true
+    }
+    return false
+}
+
 
 const getCommentsByVideoId = async (user: JwtPayload, payload: ICommentsByVideoId) => {
+    const isAuthorized = await checkAccess(user, payload.videoId)
+    if (!isAuthorized) {
+        throw new ApiError(httpStatus.UNAUTHORIZED, 'Unauthorized')
+
+    }
+
     const result = await prisma.comment.findMany({
         where: {
-            ...payload
+            ...payload,
         }
     })
     return result
 }
 
 const createComment = async (user: JwtPayload, payload: ICreateComment) => {
+    const isAuthorized = await checkAccess(user, payload.videoId)
+    if (!isAuthorized) {
+        throw new ApiError(httpStatus.UNAUTHORIZED, 'Unauthorized')
+
+    }
     const result = await prisma.comment.create({
         data: {
             userId: user.userId,
@@ -23,22 +83,13 @@ const createComment = async (user: JwtPayload, payload: ICreateComment) => {
 }
 
 const updateComment = async (user: JwtPayload, payload: IUpdateComment) => {
-    const result = await prisma.comment.updateMany({
-        where: {
-            AND: [
-                {
-                    id: {
-                        equals: payload.id
-                    }
-                },
+    const comment = await prisma.comment.findUnique({ where: { id: payload.id } })
+    if (comment?.userId !== user.userId) {
+        throw new ApiError(httpStatus.UNAUTHORIZED, 'Unauthorized')
+    }
 
-                {
-                    userId: {
-                        equals: user.userId
-                    }
-                },
-            ]
-        },
+    const result = await prisma.comment.update({
+        where: { id: payload.id },
         data: {
             comment: payload.comment
         }
@@ -47,20 +98,13 @@ const updateComment = async (user: JwtPayload, payload: IUpdateComment) => {
 }
 
 const deleteComment = async (user: JwtPayload, id: string) => {
+    const comment = await prisma.comment.findUnique({ where: { id: id } })
+    if (comment?.userId !== user.userId) {
+        throw new ApiError(httpStatus.UNAUTHORIZED, 'Unauthorized')
+    }
     const result = await prisma.comment.deleteMany({
         where: {
-            AND: [
-                {
-                    id: {
-                        equals: id
-                    }
-                },
-                {
-                    userId: {
-                        equals: user.userId
-                    }
-                },
-            ]
+            id: id
         },
     })
     return result
